% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modavg_interaction.R
\name{plot_modavg_interaction}
\alias{plot_modavg_interaction}
\title{Plot model-averaged occupancy interaction effects}
\usage{
plot_modavg_interaction(
  model_selection_table,
  full_model_list,
  cov1,
  cov2,
  delta_cutoff = 2,
  grid_length = 400,
  chunk_size = 50000,
  palette = NULL,
  xlab = NULL,
  nonInt_cov = list()
)
}
\arguments{
\item{model_selection_table}{A model selection table, typically generated by \code{modSel()} from the \code{unmarked} package.
Can also be of class \code{unmarkedModSel}.}

\item{full_model_list}{A full \code{fitList} of all candidate models (not pre-filtered by AIC).}

\item{cov1}{Character. The first covariate involved in the interaction (x-axis of the plot).}

\item{cov2}{Character. The second covariate involved in the interaction (used to color lines).}

\item{delta_cutoff}{Numeric. Threshold for selecting models within a given ΔAIC (default is 2).}

\item{grid_length}{Integer. Number of points to use in the prediction grid for each covariate (default is 400).}

\item{chunk_size}{Integer. Number of rows to predict at once (useful for memory management; default is 50,000).}

\item{palette}{Character. Optional name of a color palette passed to \code{hcl.colors()}. Defaults to \code{"cividis"}.}

\item{xlab}{Character. Optional label for the x-axis. If \code{NULL}, \code{cov1} is used.}

\item{nonInt_cov}{Named list. Values for non-interacting covariates to be held constant during prediction.
If not specified, these covariates are set to their median value from the data.}
}
\value{
A \code{ggplot2} object showing the model-averaged relationship between occupancy and the interaction
between \code{cov1} and \code{cov2}, with lines colored by \code{cov2}.
}
\description{
This function creates a plot of model-averaged occupancy probabilities
across a range of two interacting covariates using a set of models selected
based on an AIC threshold.
}
\details{
Only models containing both \code{cov1} and \code{cov2} as site-level covariates will be used. The function generates
a grid of predictions across a combination of values of \code{cov1} and \code{cov2}, and computes model-averaged
occupancy probabilities weighted by AIC support.

Any other site-level covariates not specified in \code{nonInt_cov} will be set to their median values.
}
\examples{
\dontrun{
plot_modavg_interaction(
  model_selection_table = ms,
  full_model_list = mod_ls,
  cov1 = "cov1",
  cov2 = "cov2",
  delta_cutoff = 2,
  grid_length = 400,
  nonInt_cov = list(cov3 = 0, cov4 = 0)
)
}

}
