% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occupancy.R
\name{plot_occupancy}
\alias{plot_occupancy}
\title{Plot occupancy estimates against a continuous covariate}
\usage{
plot_occupancy(model, covariate, fixed_vals = NULL, xlab = NULL)
}
\arguments{
\item{model}{An object of class \code{unmarkedFit}, typically the result of fitting a model using the \code{unmarked} package.}

\item{covariate}{A character string giving the name of the site-level covariate to be plotted on the x-axis.}

\item{fixed_vals}{Optional named list of fixed values for other site-level covariates not being plotted. If not provided, they default to 0.}

\item{xlab}{Optional custom label for the x-axis. If not specified, defaults to the covariate name with " (standardised)" appended.}
}
\value{
A \code{ggplot} object showing the predicted occupancy probabilities with confidence intervals.
}
\description{
This function generates a plot of occupancy probabilities predicted by a fitted
\code{unmarkedFit} model, as a function of a selected site-level covariate.
A ribbon shows the 95\% confidence interval around the predicted values.
}
