% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_raster.R
\name{standardise_raster}
\alias{standardise_raster}
\title{Standardise a raster Layer based on Covariate Data}
\usage{
standardise_raster(
  raster_layer,
  cov_data,
  cov_name,
  return_class = c("same", "terra", "raster")
)
}
\arguments{
\item{raster_layer}{A raster object to standardise. Must be a \code{SpatRaster} (from \code{terra}) or \code{RasterLayer} (from \code{raster}).}

\item{cov_data}{A data frame containing the covariate data used to compute mean and standard deviation.}

\item{cov_name}{A character string specifying the name of the covariate column in \code{cov_data}.}

\item{return_class}{A character string indicating the class of the returned raster: \code{"same"} (default; same as input),
\code{"terra"} to return a \code{SpatRaster}, or \code{"raster"} to return a \code{RasterLayer}.}
}
\value{
A standardized raster object of the class specified by \code{return_class}.
}
\description{
This function standardises a raster layer by subtracting the mean and dividing by the standard deviation
of a specified covariate from a provided data frame. It supports raster objects from either the
\code{raster} or \code{terra} packages and can return the result as either class.
}
\examples{
\dontrun{
library(terra)
r <- rast(system.file("ex/logo.tif", package="terra"))
cov_df <- data.frame(elevation = runif(100, 100, 500))
std_r <- standardise_raster(r, cov_df, "elevation")
plot(std_r)
}

}
